refactor: #12 refactor converting logic
