+++ src/controllers/mod.rs
@@ -2,4 +2,4 @@ pub mod root;
 pub mod dice;
 pub mod sleepy;
 pub mod metrics;
-pub mod make_rgb;
+pub mod convert;
